requirements.txt
# Core dependencies (required)
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
customtkinter>=5.2.0
scikit-learn>=1.3.0
requests>=2.31.0
tqdm>=4.65.0
tkinterdnd2>=0.3.0

# PyTorch (CPU version by default)
torch>=2.0.0
torchvision>=0.15.0

# AI/ML libraries (optional - for advanced features)
# Uncomment these if you want full AI functionality:
# transformers>=4.30.0
# diffusers>=0.18.0
# accelerate>=0.20.0
# rembg>=2.0.0

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
